requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
asgiref=3.3.1=pypi_0
backcall=0.2.0=py_0
ca-certificates=2020.10.14=0
certifi=2020.6.20=pyhd3eb1b0_3
cycler=0.10.0=pypi_0
decorator=4.4.2=py_0
django=3.1.3=pypi_0
django-chartjs=2.2.1=pypi_0
django-crispy-forms=1.9.2=pypi_0
django-menus=1.1.2=pypi_0
djangorestframework=3.12.2=pypi_0

ncurses=6.2=he6710b0_1


openssl=1.1.1h=h7b6447c_0
parso=0.7.0=py_0
pexpect=4.8.0=py38_0
pickleshare=0.7.5=py38_1000
pip=20.2.4=py38_0
prompt-toolkit=3.0.8=py_0
ptyprocess=0.6.0=py38_0
pygments=2.7.2=pyhd3eb1b0_0
pyparsing=2.4.7=pypi_0
python=3.8.5=h7579374_1
python-dateutil=2.8.1=py_0
pytz=2020.4=pypi_0
pywavelets=1.1.1=pypi_0
pyzmq=19.0.2=py38he6710b0_1
readline=8.0=h7b6447c_0
setuptools=50.3.0=py38hb0f4dca_1
six=1.15.0=py_0
sqlite=3.33.0=h62c20be_0
sqlparse=0.4.1=pypi_0
tifffile=2020.10.1=pypi_0
tk=8.6.10=hbc83047_0
tornado=6.0.4=py38h7b6447c_1
traitlets=5.0.5=py_0
wcwidth=0.2.5=py_0
wheel=0.35.1=py_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.3=he6710b0_3
zlib=1.2.11=h7b6447c_3
